refactor(mysql): centralize TLS file validation
